+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+### Changed
+
+- MySQL: Add an optional `charset_collation` to the connector, as PyMySQL >=1.1.0 always runs a `SET NAMES` on connection,
+  which breaks on servers using a non-default collation
+
+### Fix
+
+- MySQL: Allow dict parameters to be used with PyMySQL 1.1.1
+- MySQL: Use a regular PyMySQL Cursor rather than a DictCursor when pandas 2.x is used
+
 ## [6.1.2] 2024-04-18
 
 ### Fixed